PB-30b: more refactoring
