+++ application/next-frontend/src/api/utils.ts
@@ -1,13 +1,4 @@
-import { useRouter } from 'next/router'
-import useSWR from 'swr'
 import { clientsideApiUri } from './endpoints'
-
-export type FetcherError = {
-    statusCode: number
-    info: {
-        msg: string
-    }
-}
 
 const getCookie = (name: string) => {
     if (typeof window !== 'undefined') {
@@ -15,20 +6,6 @@ const getCookie = (name: string) => {
         const parts = value.split(`; ${name}=`)
         if (parts && parts.length === 2) return parts.pop()?.split(';')?.shift()
     }
-}
-
-export const useAuthedSWR = <Data>(endpoint: string) => {
-    const { get } = mutater()
-    const { data, isLoading, error, mutate } = useSWR<Data, FetcherError>(endpoint, get)
-    const router = useRouter()
-
-    if (error) {
-        if (error.statusCode === 401 || error.statusCode == 403) {
-            router.push('/login')
-        }
-    }
-
-    return { data, isLoading, error, mutate }
 }
 
 const configureHeaders = () => {
@@ -45,17 +22,15 @@ const configureHeaders = () => {
     return headers
 }
 
-const fetchData = async (endpoint: string, method: 'GET' | 'POST' | 'PUT' | 'DELETE', body?: object) => {
+const fetchData = async <Data>(endpoint: string, method: 'GET' | 'POST' | 'PUT' | 'DELETE', body?: object) => {
     const options: RequestInit = {
         method: method,
         headers: configureHeaders(),
     }
     if (body) options.body = JSON.stringify(body)
 
-    return await fetch(clientsideApiUri + endpoint, options)
-}
+    const res = await fetch(clientsideApiUri + endpoint, options)
 
-const handleHttpResponse = async <Data>(res: Response) => {
     if (!res.ok) {
         const info = await res.json()
         const err = { statusCode: res.status, info }
@@ -64,28 +39,24 @@ const handleHttpResponse = async <Data>(res: Response) => {
     return res.json() as Data
 }
 
-export const mutater = () => {
+export const apiRequestHelper = () => {
     /**
      * Should not be used directly, but through the useAuthedSWR hook
      */
     const get = async <Data>(endpoint: string) => {
-        const res = await fetchData(endpoint, 'GET')
-        return handleHttpResponse<Data>(res)
+        return await fetchData<Data>(endpoint, 'GET')
     }
 
     const put = async (endpoint: string, body: object) => {
-        const res = await fetchData(endpoint, 'PUT', body)
-        return handleHttpResponse(res)
+        return await fetchData(endpoint, 'PUT', body)
     }
 
     const post = async (endpoint: string, body: object) => {
-        const res = await fetchData(endpoint, 'POST', body)
-        return handleHttpResponse(res)
+        return await fetchData(endpoint, 'POST', body)
     }
 
     const del = async (endpoint: string) => {
-        const res = await fetchData(endpoint, 'DELETE')
-        return await handleHttpResponse(res)
+        return await fetchData(endpoint, 'DELETE')
     }
 
     return { get, put, post, del }
